Fix circular dependencies for comand
