+++ project/api/views/endpoints/mvt_stv.py
@@ -22,7 +22,49 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 from django.http import HttpResponse
 from django.db import connection
 from api.models import MVTLayers
-from api.management.commands.generate_mvt import stv_mvt_geom_query
+
+
+def mvt_geom_params(zoom):
+    """ Simplification for territory field """
+    # For WebMercator (3857) X coordinate bounds are ±20037508.3427892 meters
+    # For SRID 4326 X coordinated bounds are ±180 degrees
+    # resolution = (xmax - xmin) or (xmax * 2)
+    # It is 5-10 times faster work with SRID 4326,
+    # We will apply ST_Simplify before ST_Transform
+    resolution = 360
+    # https://postgis.net/docs/ST_AsMVT.html
+    # tile extent in screen space as defined by the specification
+    extent = 4096
+
+    # Find safe tolerance for ST_Simplfy
+    tolerance = (float(resolution) / 2 ** zoom) / float(extent)
+    # Apply additional simplification for distant zoom levels
+    tolerance_multiplier = 1 if zoom > 5 else 2.2 - 0.2 * zoom
+    simplification = tolerance * tolerance_multiplier
+
+    return """
+        ST_SnapToGrid(
+            ST_Transform(
+                ST_Simplify(territory, {})
+                , 3857
+            ), 1
+        )
+    """.format(
+        simplification
+    )
+
+
+def stv_mvt_geom_query(zoom):
+    """ query for stv mvt tile """
+    return """
+        SELECT
+            id, start_date, end_date, "references", entity_id, wikidata_id, color, admin_level
+            , ST_AsMVTGeom({}, TileBBox(%(zoom)s, %(x_coor)s, %(y_coor)s)) as territory
+        FROM view_stvmap
+        WHERE ST_Intersects(territory, TileBBox(%(zoom)s, %(x_coor)s, %(y_coor)s, 4326))
+    """.format(
+        mvt_geom_params(zoom)
+    )
 
 
 def parse_ints(arr):
@@ -57,7 +99,9 @@ def mvt_stv(request, zoom, x_coor, y_coor):
                 """
                 SELECT ST_AsMVT(a, 'stv_admin') AS tile
                 FROM ({} AND ({})) AS a
-                """.format(stv_mvt_geom_query(zoom), " OR ".join(where)),
+                """.format(
+                    stv_mvt_geom_query(zoom), " OR ".join(where)
+                ),
                 {"zoom": zoom, "x_coor": x_coor, "y_coor": y_coor,},
             )
             tile = bytes(cursor.fetchone()[0])
